Fix grunfile.js
Version was not properly being injected before. Currently that is serving as the initial declaration of Polymer if you are running platform-less. For instance the current master branch is broken if you are running without platform.js because of this. It looks like the 0.4.0 release tag does have the version (so it works), but I am not sure how it was added?